+++ NEWS.md
@@ -9,6 +9,7 @@ After we release v1 of the package, we will start following [semantic versioning
 ITensor v0.2.2 Release Notes
 ==============================
 
+- Make Index non-broadcastable so you can do: `i = Index(2); i .^ (0, 1, 2)` (PR #689).
 - Add interface `contract([A, [B, C]])` for recursively contracting a tensor network tree, equivalent to `contract([A, B, C]; sequence=[1, [2, 3]])` (PR #686).
 - Allow plain integer tags, such as `Index(2, "1")` (PR #686).